requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2023.1
Django==4.2
django-crispy-forms==2.1
django-environ==0.11.2
django-filter==23.5
django-localflavor==4.0
Pillow==10.1.0
python-stdnum==1.19
sqlparse==0.4.4
typing_extensions==4.9.0
